+++ src/Simd/SimdFont.hpp
@@ -89,7 +89,7 @@ namespace Simd
         */
         bool Resize(size_t height)
         {
-            if (height == _currentSize.y)
+            if (height == static_cast<std::size_t>(_currentSize.y))
                 return true;
 
             if (height < 4u || height > (size_t)_originalSize.y*4)
