+++ src/Simd/SimdSse41ImageLoadPng.cpp
@@ -518,7 +518,7 @@ namespace Simd
             if (s >= 16) return -1; // invalid code!
             // code size is s, so:
             b = (k >> (16 - s)) - z->firstcode[s] + z->firstsymbol[s];
-            if (b >= sizeof(z->size)) return -1; // some data was corrupt somewhere!
+            if (static_cast<std::size_t>(b) >= sizeof(z->size)) return -1; // some data was corrupt somewhere!
             if (z->size[b] != s) return -1;  // was originally an assert, but report failure instead.
             a->code_buffer >>= s;
             a->num_bits -= s;
@@ -1751,7 +1751,7 @@ namespace Simd
                 }
                 size_t stride = 4 * x;
                 _image.Recreate(x, y, (Image::Format)_param.format);
-                if (x < A)
+                if (static_cast<std::size_t>(x) < A)
                 {
                     switch (_param.format)
                     {
